update kts, renstra nilai; penambahan komponen ckeditor, choice & banner;

+++ src/Page/IndikatorRenstra/IndikatorRenstraPage.jsx
@@ -6,6 +6,8 @@ import Navbar from "@/Components/Navbar";
 import ChangeLevelModal from "@/Components/ChangeLevelModal";
 import { useContent } from "@/Providers/ContentProvider";
 import { BsPlus } from "react-icons/bs";
+import { useAuth } from "@/Providers/AuthProvider";
+import { isEmpty } from "@/Common/Utils";
 
 const IndikatorRenstraPage = () => {
   const [confirmDelete, setConfirmDelete] = useState(false);
@@ -13,6 +15,7 @@ const IndikatorRenstraPage = () => {
   const [modalOpen, setModalOpen] = useState(false);
   const [modalMode, setModalMode] = useState("new");
   const [selectedRow, setSelectedRow] = useState(null);
+  const {getValidToken} = useAuth()
 
   const openNew = () => {
     setModalMode("new");
@@ -27,10 +30,15 @@ const IndikatorRenstraPage = () => {
   };
 
   const deleteData = async () => {
+    if(isEmpty(getValidToken())) return;
+
     const res = await fetch(
       `http://localhost:3000/standarrenstra/${selectedRow.Uuid}`,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${getValidToken()}`
+        }
       }
     );
     const data = await res.json();
@@ -46,23 +54,16 @@ const IndikatorRenstraPage = () => {
     tableRef.current?.reload?.({ resetPage: true });
   };
 
-  const { level, setLevel, openChangeLevel, setOpenChangeLevel } = useContent();
+  const { level, listLevel, setLevel, openChangeLevel, setOpenChangeLevel } = useContent();
 
   return (
     <>
       <Navbar
-        userName="John Doe"
-        userLevel={level}
-        years={[]}
-        activeYear={null}
-        positionYear={null}
-        onPositionChange={() => {}}
-        onChangeLevelClick={() => setOpenChangeLevel(true)}
         renderChangeLevelModal={() => (
           <ChangeLevelModal
             open={openChangeLevel}
             onClose={() => setOpenChangeLevel(false)}
-            levels={[]}
+            levels={listLevel}
             currentLevel={level}
             onSubmit={(val) => {
               setLevel(val);
@@ -75,57 +76,58 @@ const IndikatorRenstraPage = () => {
         <h2 className="text-lg font-semibold mb-4">Indikator Renstra</h2>
 
         <div className="border rounded mt-3">
-          <RemoteTable
-            ref={tableRef}
-            endpoint="http://localhost:3000/indikatorrenstras"
-            mode="sse"
-            onError={(err) => {
-              console.error("TABLE ERROR:", err);
-            }}
-            renderAddAction={
-              <button
-                className="px-3 py-2 bg-purple-600 text-white rounded-lg"
-                onClick={() => openNew()}
-              >
-                <BsPlus />
-              </button>
-            }
-            listcolumns={[
-              { key: "Tahun", label: "Tahun", searchable: true },
-              { key: "Standar", label: "Standar Standar", searchable: true },
-              {
-                key: "Indikator",
-                label: "Indikator Standar",
-                searchable: true,
-              },
-            ]}
-            renderAction={({ row, close }) => (
-              <>
+            <RemoteTable
+              ref={tableRef}
+              endpoint="http://localhost:3000/indikatorrenstras"
+              mode="sse"
+              token={getValidToken()}
+              onError={(err) => {
+                console.error("TABLE ERROR:", err);
+              }}
+              renderAddAction={
                 <button
-                  className="block w-full px-3 py-2 text-sm hover:bg-gray-100"
-                  onClick={() => {
-                    console.log("edit", row);
-                    openEdit(row);
-                    close();
-                  }}
+                  className="px-3 py-2 bg-purple-600 text-white rounded-lg"
+                  onClick={() => openNew()}
                 >
-                  Edit
+                  <BsPlus />
                 </button>
+              }
+              listcolumns={[
+                { key: "Tahun", label: "Tahun", searchable: true },
+                { key: "Standar", label: "Standar Standar", searchable: true },
+                {
+                  key: "Indikator",
+                  label: "Indikator Standar",
+                  searchable: true,
+                },
+              ]}
+              renderAction={({ row, close }) => (
+                <>
+                  <button
+                    className="block w-full px-3 py-2 text-sm hover:bg-gray-100"
+                    onClick={() => {
+                      console.log("edit", row);
+                      openEdit(row);
+                      close();
+                    }}
+                  >
+                    Edit
+                  </button>
 
-                <button
-                  className="block w-full px-3 py-2 text-sm text-red-600 hover:bg-gray-100"
-                  onClick={() => {
-                    console.log("hapus", row);
-                    setSelectedRow(row);
-                    setConfirmDelete(true);
-                    close();
-                  }}
-                >
-                  Hapus
-                </button>
-              </>
-            )}
-          />
+                  <button
+                    className="block w-full px-3 py-2 text-sm text-red-600 hover:bg-gray-100"
+                    onClick={() => {
+                      console.log("hapus", row);
+                      setSelectedRow(row);
+                      setConfirmDelete(true);
+                      close();
+                    }}
+                  >
+                    Hapus
+                  </button>
+                </>
+              )}
+            />
         </div>
 
         <IndikatorRenstraFormModal
