+++ src/Page/JenisFile/JenisFilePage.jsx
@@ -7,8 +7,8 @@ import { useContent } from "@/Providers/ContentProvider";
 import Navbar from "@/Components/Navbar";
 import ChangeLevelModal from "@/Components/ChangeLevelModal";
 import { BsPlus } from "react-icons/bs";
-
-// const cn = (...classes) => classes.filter(Boolean).join(" ");
+import { useAuth } from "@/Providers/AuthProvider";
+import { isEmpty } from "@/Common/Utils";
 
 const JenisFilePage = () => {
   const [confirmDelete, setConfirmDelete] = useState(false);
@@ -17,6 +17,7 @@ const JenisFilePage = () => {
   const [modalMode, setModalMode] = useState("new");
   const [selectedRow, setSelectedRow] = useState(null);
   const { addToast } = useToast();
+  const {getValidToken} = useAuth()
 
   const openNew = () => {
     setModalMode("new");
@@ -31,10 +32,15 @@ const JenisFilePage = () => {
   };
 
   const deleteData = async () => {
+    if(isEmpty(getValidToken())) return;
+
     const res = await fetch(
       `http://localhost:3000/jenisfile/${selectedRow.UUID}`,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${getValidToken()}`
+        }
       }
     );
     const data = await res.json();
@@ -50,23 +56,16 @@ const JenisFilePage = () => {
     tableRef.current?.reload?.({ resetPage: true });
   };
 
-  const { level, setLevel, openChangeLevel, setOpenChangeLevel } = useContent();
+  const { level, listLevel, setLevel, openChangeLevel, setOpenChangeLevel } = useContent();
 
   return (
     <>
       <Navbar
-        userName="John Doe"
-        userLevel={level}
-        years={[]}
-        activeYear={null}
-        positionYear={null}
-        onPositionChange={() => {}}
-        onChangeLevelClick={() => setOpenChangeLevel(true)}
         renderChangeLevelModal={() => (
           <ChangeLevelModal
             open={openChangeLevel}
             onClose={() => setOpenChangeLevel(false)}
-            levels={[]}
+            levels={listLevel}
             currentLevel={level}
             onSubmit={(val) => {
               setLevel(val);
@@ -79,48 +78,49 @@ const JenisFilePage = () => {
         <h2 className="text-lg font-semibold mb-4">Jenis File Audit</h2>
 
         <div className="border rounded">
-          <RemoteTable
-            ref={tableRef}
-            endpoint="http://localhost:3000/jenisfiles"
-            mode="paging"
-            renderAddAction={
-              <button
-                className="px-3 py-2 bg-purple-600 text-white rounded-lg"
-                onClick={() => openNew()}
-              >
-                <BsPlus />
-              </button>
-            }
-            listcolumns={[
-              { key: "Nama", label: "Jenis File", searchable: true },
-            ]}
-            renderAction={({ row, close }) => (
-              <>
+            <RemoteTable
+              ref={tableRef}
+              endpoint="http://localhost:3000/jenisfiles"
+              mode="paging"
+              token={getValidToken()}
+              renderAddAction={
                 <button
-                  className="block w-full px-3 py-2 text-sm hover:bg-gray-100"
-                  onClick={() => {
-                    console.log("edit", row);
-                    openEdit(row);
-                    close();
-                  }}
+                  className="px-3 py-2 bg-purple-600 text-white rounded-lg"
+                  onClick={() => openNew()}
                 >
-                  Edit
+                  <BsPlus />
                 </button>
+              }
+              listcolumns={[
+                { key: "Nama", label: "Jenis File", searchable: true },
+              ]}
+              renderAction={({ row, close }) => (
+                <>
+                  <button
+                    className="block w-full px-3 py-2 text-sm hover:bg-gray-100"
+                    onClick={() => {
+                      console.log("edit", row);
+                      openEdit(row);
+                      close();
+                    }}
+                  >
+                    Edit
+                  </button>
 
-                <button
-                  className="block w-full px-3 py-2 text-sm text-red-600 hover:bg-gray-100"
-                  onClick={() => {
-                    console.log("hapus", row);
-                    setSelectedRow(row);
-                    setConfirmDelete(true);
-                    close();
-                  }}
-                >
-                  Hapus
-                </button>
-              </>
-            )}
-          />
+                  <button
+                    className="block w-full px-3 py-2 text-sm text-red-600 hover:bg-gray-100"
+                    onClick={() => {
+                      console.log("hapus", row);
+                      setSelectedRow(row);
+                      setConfirmDelete(true);
+                      close();
+                    }}
+                  >
+                    Hapus
+                  </button>
+                </>
+              )}
+            /> 
         </div>
 
         <JenisFileFormModal
